Newmark time integration and xml based output for paraview
git-svn-id: svn+ssh://svn.code.sf.net/p/nuto/code/trunk@598 b5606e62-d4ae-482b-8a79-2ac8f2527b1c

+++ src/nuto/math/SparseMatrixCSRSymmetric.cpp
@@ -337,8 +337,8 @@ FullMatrix<double> SparseMatrixCSRSymmetric<double>::operator* (const FullMatrix
         {
             for (int thisPos = this->mRowIndex[thisRow]; thisPos < this->mRowIndex[thisRow + 1]; thisPos++)
             {
-                int thisColumn = this->mColumns[thisPos];
-                int thisValue = this->mValues[thisPos];
+            	int thisColumn = this->mColumns[thisPos];
+                double thisValue = this->mValues[thisPos];
                 resultValues[thisRow] += thisValue * matrixValues[thisColumn];
                 if (thisRow != thisColumn)
                 {
